@@ -1,25 +1,32 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
+plugins {
+    id 'com.android.application'
+    id 'kotlin-android'
+    id 'kotlin-android-extensions'
+    id 'kotlin-kapt'
+}
 
 android {
+
     signingConfigs {
         release {
-            storeFile file('gamercollection-keystore.jks')
-            storePassword 'Klabe-keystore'
-            keyAlias 'gamercollection'
-            keyPassword 'Klabe-gamercollection'
+            if (project.hasProperty('GAMER_COLLECTION_STORE_FILE')) {
+                storeFile file(GAMER_COLLECTION_STORE_FILE)
+                storePassword GAMER_COLLECTION_STORE_PASSWORD
+                keyAlias GAMER_COLLECTION_KEY_ALIAS
+                keyPassword GAMER_COLLECTION_KEY_PASSWORD
+            }
         }
     }
-    compileSdkVersion 29
-    buildToolsVersion "29.0.2"
+
+    compileSdkVersion 30
+    buildToolsVersion "30.0.2"
 
     defaultConfig {
         applicationId "es.upsa.mimo.gamercollection"
-        minSdkVersion 22
-        targetSdkVersion 29
-        versionCode 2
-        versionName "1.0.1"
+        minSdkVersion 23
+        targetSdkVersion 30
+        versionCode 3
+        versionName "2.0.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
