@@ -1,23 +1,28 @@
 package es.upsa.mimo.gamercollection.adapters
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
+import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
-import es.upsa.mimo.gamercollection.R
-import es.upsa.mimo.gamercollection.models.SongResponse
-import kotlinx.android.synthetic.main.song_item.view.*
+import es.upsa.mimo.gamercollection.databinding.SongItemBinding
+import es.upsa.mimo.gamercollection.models.responses.SongResponse
+import es.upsa.mimo.gamercollection.utils.Constants
 
 class SongsAdapter(
-    var songs: List<SongResponse>,
-    var editable: Boolean,
+    private var songs: List<SongResponse>,
+    private var editable: Boolean,
     private var onItemClickListener: OnItemClickListener
-): RecyclerView.Adapter<SongsAdapter.SongsViewHolder?>() {
+) : RecyclerView.Adapter<SongsAdapter.SongsViewHolder?>() {
 
+    //region Lifecycle methods
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SongsViewHolder {
-
-        val itemView: View = LayoutInflater.from(parent.context).inflate(R.layout.song_item, parent, false)
-        return SongsViewHolder(itemView)
+        return SongsViewHolder(
+            SongItemBinding.inflate(
+                LayoutInflater.from(parent.context),
+                parent,
+                false
+            )
+        )
     }
 
     override fun getItemCount(): Int {
@@ -25,20 +30,54 @@ class SongsAdapter(
     }
 
     override fun onBindViewHolder(holder: SongsViewHolder, position: Int) {
+        holder.bind(songs[position])
+    }
+    //endregion
 
-        val song = songs[position]
-        holder.itemView.text_view_name.text = song.name
-        holder.itemView.text_view_singer.text = song.singer
-        holder.itemView.text_view_url.text = song.url
-        holder.itemView.image_view_remove.visibility = if (editable) View.VISIBLE else View.GONE
-        holder.itemView.image_view_remove.setOnClickListener {
-            onItemClickListener.onItemClick(song.id)
+    //region Public methods
+    fun setSongs(newSongs: List<SongResponse>) {
+
+        this.songs = newSongs
+        notifyDataSetChanged()
+    }
+
+    fun setEditable(editable: Boolean) {
+
+        this.editable = editable
+        notifyDataSetChanged()
+    }
+    //endregion
+
+    inner class SongsViewHolder(val binding: SongItemBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+
+        fun bind(_song: SongResponse) {
+            with(binding) {
+
+                darkMode = Constants.isDarkMode(root.context)
+                song = _song
+                editable = this@SongsAdapter.editable
+                onItemClickListener = this@SongsAdapter.onItemClickListener
+                executePendingBindings()
+            }
         }
     }
+}
 
-    interface OnItemClickListener {
-        fun onItemClick(songId: Int)
+@BindingAdapter("songs")
+fun setRecyclerViewSongs(recyclerView: RecyclerView?, songs: List<SongResponse>?) {
+
+    val adapter = recyclerView?.adapter
+    if (adapter is SongsAdapter && songs != null) {
+        adapter.setSongs(songs)
     }
+}
 
-    class SongsViewHolder(itemView: View): RecyclerView.ViewHolder(itemView)
+@BindingAdapter("editable")
+fun setRecyclerViewEditable(recyclerView: RecyclerView?, editable: Boolean?) {
+
+    val adapter = recyclerView?.adapter
+    if (adapter is SongsAdapter && editable != null) {
+        adapter.setEditable(editable)
+    }
 }
