@@ -6,60 +6,74 @@ import android.content.Intent;
 import android.content.res.Configuration;
 import android.os.Build;
 import android.os.Bundle;
+
+import androidx.lifecycle.ViewModelProvider;
+
 import java.util.Locale;
+
 import es.upsa.mimo.gamercollection.R;
-import es.upsa.mimo.gamercollection.activities.base.BaseActivity;
-import es.upsa.mimo.gamercollection.utils.Constants;
-import es.upsa.mimo.gamercollection.utils.SharedPreferencesHandler;
+import es.upsa.mimo.gamercollection.base.BaseActivity;
+import es.upsa.mimo.gamercollection.utils.Notifications;
+import es.upsa.mimo.gamercollection.viewmodelfactories.LandingViewModelFactory;
+import es.upsa.mimo.gamercollection.viewmodels.LandingViewModel;
 
 public class LandingActivity extends BaseActivity {
 
+    //region Private properties
+    private LandingViewModel viewModel;
+    //endregion
+
+    //region Lifecycle methods
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        initializeUI();
+    }
+    //endregion
 
+    //region Private methods
+    private void initializeUI() {
+
+        viewModel = new ViewModelProvider(this, new LandingViewModelFactory(getApplication())).get(LandingViewModel.class);
+        setupBindings();
+
+        configLanguage();
         createNotificationChannel();
 
-        SharedPreferencesHandler sharedPrefHandler = new SharedPreferencesHandler(this);
+        viewModel.checkVersion();
+        viewModel.checkTheme();
+    }
 
-        String language = sharedPrefHandler.getLanguage();
+    private void setupBindings() {
+
+        viewModel.getLandingClassToStart().observe(this, cls -> {
+
+            Intent intent = new Intent(this, cls);
+            startActivity(intent);
+        });
+    }
+
+    private void configLanguage() {
+
+        String language = viewModel.getLanguage();
         Configuration conf = getResources().getConfiguration();
         conf.setLocale(new Locale(language));
         getResources().updateConfiguration(conf, getResources().getDisplayMetrics());
-
-        Class<?> cls;
-        Long initTime = System.currentTimeMillis() / 1000;
-        if (sharedPrefHandler.isLoggedIn()) {
-            cls = MainActivity.class;
-        } else {
-            cls = LoginActivity.class;
-        }
-        Long finalTime = System.currentTimeMillis() / 1000;
-        long taskTime = finalTime - initTime;
-        long time = Math.max(0, 1000 - taskTime);
-
-        try {
-            Thread.sleep(time);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        Intent intent = new Intent(this, cls);
-        startActivity(intent);
     }
-
-    //MARK: - Private functions
 
     private void createNotificationChannel() {
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            String name = getString(R.string.CHANNEL_NAME);
-            String description = getString(R.string.CHANNEL_DESCRIPTION);
+
+            String name = getString(R.string.channel_name);
+            String description = getString(R.string.channel_description);
             int importance = NotificationManager.IMPORTANCE_DEFAULT;
-            NotificationChannel channel = new NotificationChannel(Constants.channelId, name, importance);
+
+            NotificationChannel channel = new NotificationChannel(Notifications.CHANNEL_ID, name, importance);
             channel.setDescription(description);
             NotificationManager notificationManager = getSystemService(NotificationManager.class);
             if (notificationManager != null) notificationManager.createNotificationChannel(channel);
         }
     }
+    //endregion
 }
