@@ -1,87 +1,64 @@
 package es.upsa.mimo.gamercollection.fragments
 
-import android.app.AlertDialog
 import android.app.Notification
 import android.app.PendingIntent
 import android.content.Intent
 import android.os.Bundle
 import android.view.*
-import android.widget.LinearLayout
-import android.widget.NumberPicker
+import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
+import androidx.databinding.ObservableField
 import androidx.fragment.app.FragmentTransaction
+import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.sqlite.db.SimpleSQLiteQuery
+import androidx.recyclerview.widget.RecyclerView
 import es.upsa.mimo.gamercollection.R
 import es.upsa.mimo.gamercollection.activities.GameDetailActivity
 import es.upsa.mimo.gamercollection.adapters.GamesAdapter
-import es.upsa.mimo.gamercollection.fragments.base.BaseFragment
-import es.upsa.mimo.gamercollection.fragments.popups.OnFiltersSelected
+import es.upsa.mimo.gamercollection.adapters.OnFiltersSelected
+import es.upsa.mimo.gamercollection.adapters.OnItemClickListener
+import es.upsa.mimo.gamercollection.base.BindingFragment
+import es.upsa.mimo.gamercollection.databinding.FragmentGamesBinding
 import es.upsa.mimo.gamercollection.fragments.popups.PopupFilterDialogFragment
 import es.upsa.mimo.gamercollection.models.FilterModel
-import es.upsa.mimo.gamercollection.models.GameResponse
-import es.upsa.mimo.gamercollection.network.apiClient.GameAPIClient
-import es.upsa.mimo.gamercollection.persistence.repositories.GameRepository
-import es.upsa.mimo.gamercollection.persistence.repositories.PlatformRepository
-import es.upsa.mimo.gamercollection.persistence.repositories.StateRepository
+import es.upsa.mimo.gamercollection.models.responses.GameResponse
 import es.upsa.mimo.gamercollection.utils.Constants
-import es.upsa.mimo.gamercollection.utils.SharedPreferencesHandler
-import kotlinx.android.synthetic.main.fragment_games.*
+import es.upsa.mimo.gamercollection.utils.Notifications
+import es.upsa.mimo.gamercollection.utils.State
+import es.upsa.mimo.gamercollection.viewmodelfactories.GamesViewModelFactory
+import es.upsa.mimo.gamercollection.viewmodels.GamesViewModel
 import kotlinx.android.synthetic.main.state_button.view.*
 import java.util.*
 import kotlin.collections.ArrayList
 
-class GamesFragment : BaseFragment(), GamesAdapter.OnItemClickListener, OnFiltersSelected {
+class GamesFragment : BindingFragment<FragmentGamesBinding>(), OnItemClickListener,
+    OnFiltersSelected {
 
-    private lateinit var sharedPrefHandler: SharedPreferencesHandler
-    private lateinit var gameAPIClient: GameAPIClient
-    private lateinit var gameRepository: GameRepository
-    private lateinit var platformRepository: PlatformRepository
-    private lateinit var stateRepository: StateRepository
+    //region Private properties
+    private lateinit var viewModel: GamesViewModel
+    private lateinit var gamesAdapter: GamesAdapter
     private var menu: Menu? = null
-    private lateinit var sortingKeys: Array<String>
-    private var sortingValues = arrayOf("")
-    private var state: String? = null
-    private lateinit var sortKey: String
-    private var sortAscending = true
-    private var currentFilters: FilterModel? = null
+    private val scrollPosition = ObservableField<ScrollPosition>()
+    //endregion
 
+    //region Lifecycle methods
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         setHasOptionsMenu(true)
-        return inflater.inflate(R.layout.fragment_games, container, false)
+        return super.onCreateView(inflater, container, savedInstanceState)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        sharedPrefHandler = SharedPreferencesHandler(context)
-        gameAPIClient = GameAPIClient(resources, sharedPrefHandler)
-        gameRepository = GameRepository(requireContext())
-        platformRepository = PlatformRepository(requireContext())
-        stateRepository = StateRepository(requireContext())
-        sortingKeys = resources.getStringArray(R.array.sorting_keys_ids)
-        sortingValues = resources.getStringArray(R.array.sorting_keys)
-        sortKey = sharedPrefHandler.getSortingKey()
-
         initializeUI()
     }
 
     override fun onResume() {
         super.onResume()
-
-        val games = getContent(state, sortKey, sortAscending, currentFilters)
-        setGamesCount(games)
-        val today = Constants.stringToDate(Constants.dateToString(Date(), sharedPrefHandler), sharedPrefHandler)
-        val gamesToNotify = ArrayList<GameResponse>()
-        for (game in games) {
-            if (game.releaseDate == today)
-                gamesToNotify.add(game)
-        }
-        if (gamesToNotify.isNotEmpty()) launchNotification(gamesToNotify)
+        viewModel.getGames()
     }
 
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
@@ -94,260 +71,200 @@ class GamesFragment : BaseFragment(), GamesAdapter.OnItemClickListener, OnFilter
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
 
-        when(item.itemId) {
+        when (item.itemId) {
             R.id.action_synchronize -> {
+
                 openSyncPopup()
                 return true
             }
             R.id.action_filter -> {
+
                 filter()
                 return true
             }
             R.id.action_filter_on -> {
+
                 filter()
                 return true
             }
+            R.id.action_sort_on -> {
+
+                viewModel.sortGames(requireContext(), resources)
+                return true
+            }
             R.id.action_add -> {
-                add()
+
+                launchActivity(GameDetailActivity::class.java)
                 return true
             }
         }
         return super.onOptionsItemSelected(item)
     }
+    //endregion
 
-    override fun onItemClick(gameId: Int) {
-        
-        val params = mapOf("gameId" to gameId)
+    //region Interface methods
+    override fun onItemClick(id: Int) {
+
+        val params = mapOf(Constants.GAME_ID to id, Constants.IS_RAWG_GAME to false)
         launchActivityWithExtras(GameDetailActivity::class.java, params)
+    }
+
+    override fun onSubItemClick(id: Int) {
+    }
+
+    override fun onLoadMoreItemsClick() {
     }
 
     override fun filter(filters: FilterModel?) {
 
-        currentFilters = filters
-        val games = getContent(state, sortKey, sortAscending, currentFilters)
-        setGamesCount(games)
-    }
-
-    //MARK: - Private functions
-
-    private fun initializeUI() {
-
-        button_sort.setOnClickListener { sort() }
-
-        button_pending.setOnClickListener {
-            it.isSelected = !it.isSelected
-            button_in_progress.isSelected = false
-            button_finished.isSelected = false
-            swipe_refresh_layout.isEnabled = !it.isSelected && sharedPrefHandler.getSwipeRefresh()
-            state = if (it.isSelected) Constants.pending else null
-            getContent(state, sortKey, sortAscending, currentFilters)
-        }
-        button_in_progress.setOnClickListener {
-
-            button_pending.isSelected = false
-            it.isSelected = !it.isSelected
-            button_finished.isSelected = false
-            swipe_refresh_layout.isEnabled = !it.isSelected && sharedPrefHandler.getSwipeRefresh()
-            state = if (it.isSelected) Constants.inProgress else null
-            getContent(state, sortKey, sortAscending, currentFilters)
-        }
-        button_finished.setOnClickListener {
-
-            button_pending.isSelected = false
-            button_in_progress.isSelected = false
-            it.isSelected = !it.isSelected
-            swipe_refresh_layout.isEnabled = !it.isSelected && sharedPrefHandler.getSwipeRefresh()
-            state = if (it.isSelected) Constants.finished else null
-            getContent(state, sortKey, sortAscending, currentFilters)
-        }
-
-        swipe_refresh_layout.isEnabled = sharedPrefHandler.getSwipeRefresh()
-        swipe_refresh_layout.setColorSchemeResources(R.color.color3)
-        swipe_refresh_layout.setProgressBackgroundColorSchemeResource(R.color.color2)
-        swipe_refresh_layout.setOnRefreshListener {
-            loadGames()
-        }
-
-        recycler_view_games.layoutManager = LinearLayoutManager(requireContext())
-        val platforms = platformRepository.getPlatforms()
-        val states = stateRepository.getStates()
-        recycler_view_games.adapter = GamesAdapter(requireContext(), ArrayList(), platforms, states, null, this)
-    }
-
-    private fun getContent(state: String?, sortKey: String?, sortAscending: Boolean, filters: FilterModel?): List<GameResponse> {
-
-        var queryString = "SELECT * FROM Game"
-
-        var queryConditions = when(state) {
-            Constants.pending -> " WHERE state == '${Constants.pending}' AND "
-            Constants.inProgress -> " WHERE state == '${Constants.inProgress}' AND "
-            Constants.finished -> " WHERE state == '${Constants.finished}' AND "
-            else -> ""
-        }
-
-        menu?.let{
+        viewModel.filters = filters
+        menu?.let {
             it.findItem(R.id.action_filter).isVisible = filters == null
             it.findItem(R.id.action_filter_on).isVisible = filters != null
         }
+        scrollPosition.set(ScrollPosition.TOP)
+        viewModel.getGames()
+    }
+    //endregion
 
-        filters?.let {
+    //region Public methods
+    fun buttonClicked(it: View) {
 
-            if (queryConditions.isEmpty()) queryConditions += " WHERE "
+        scrollPosition.set(ScrollPosition.TOP)
+        with(binding) {
 
-            var queryPlatforms = ""
-            val platforms = filters.platforms
-            if (platforms.isNotEmpty()) {
-                queryPlatforms += "("
-                for (platform in platforms) {
-                    queryPlatforms += "platform == '${platform}' OR "
-                }
-                queryPlatforms = queryPlatforms.dropLast(4) + ") AND "
+            swipeRefreshLayout.isEnabled =
+                !it.isSelected && this@GamesFragment.viewModel.swipeRefresh
+            buttonPending.isSelected = if (it == buttonPending) !it.isSelected else false
+            buttonInProgress.isSelected = if (it == buttonInProgress) !it.isSelected else false
+            buttonFinished.isSelected = if (it == buttonFinished) !it.isSelected else false
+            val newState = when (it) {
+                buttonPending -> State.PENDING_STATE
+                buttonInProgress -> State.IN_PROGRESS_STATE
+                buttonFinished -> State.FINISHED_STATE
+                else -> null
             }
-
-            var queryGenres = ""
-            val genres = filters.genres
-            if (genres.isNotEmpty()){
-                queryGenres += "("
-                for (genre in genres) {
-                    queryGenres += "genre == '${genre}' OR "
-                }
-                queryGenres = queryGenres.dropLast(4) + ") AND "
-            }
-
-            var queryFormats = ""
-            val formats = filters.formats
-            if (formats.isNotEmpty()){
-                queryFormats += "("
-                for (format in formats) {
-                    queryFormats += "format == '${format}' OR "
-                }
-
-                queryFormats = queryFormats.dropLast(4) + ") AND "
-            }
-
-            queryConditions += queryPlatforms + queryGenres + queryFormats
-
-            queryConditions += "score >= ${filters.minScore} AND score <= ${filters.maxScore} AND "
-
-            if (filters.minReleaseDate != null) {
-                queryConditions += "releaseDate >= '${filters.minReleaseDate}' AND "
-            }
-            if (filters.maxReleaseDate != null) {
-                queryConditions += "releaseDate <= '${filters.maxReleaseDate}' AND "
-            }
-
-            if (filters.minPurchaseDate != null) {
-                queryConditions += "purchaseDate >= '${filters.minPurchaseDate}' AND "
-            }
-            if (filters.maxPurchaseDate != null) {
-                queryConditions += "purchaseDate <= '${filters.maxPurchaseDate}' AND "
-            }
-
-            queryConditions += "price >= ${filters.minPrice} AND "
-            if (filters.maxPrice > 0) {
-                queryConditions += "price <= ${filters.maxPrice} AND "
-            }
-
-            if (filters.isGoty) {
-                queryConditions += "goty == 1 AND "
-            }
-
-            if (filters.isLoaned) {
-                queryConditions += "loanedTo != null AND "
-            }
-
-            if (filters.hasSaga) {
-                queryConditions += "saga_id != -1 AND "
-            }
-
-            if (filters.hasSongs) {
-                queryConditions += "songs != '[]' AND "
-            }
+            this@GamesFragment.viewModel.state = if (it.isSelected) newState else null
         }
-        queryConditions = queryConditions.dropLast(5)
-        queryString += queryConditions
-
-        val sortParam = sortKey ?: sharedPrefHandler.getSortingKey()
-        val sortOrder = if(sortAscending) "ASC"  else "DESC"
-        queryString += " ORDER BY $sortParam $sortOrder"
-
-        val query = SimpleSQLiteQuery(queryString)
-        val games = gameRepository.getGames(query)
-
-        val adapter = recycler_view_games.adapter
-        if (adapter != null && adapter is GamesAdapter) {
-            adapter.games = games
-            adapter.notifyDataSetChanged()
-        }
-        layout_empty_list.visibility = if (games.isNotEmpty()) View.GONE else View.VISIBLE
-        swipe_refresh_layout.visibility = if (games.isNotEmpty()) View.VISIBLE else View.GONE
-
-        return games
+        viewModel.getGames()
     }
 
-    private fun setGamesCount(games: List<GameResponse>) {
+    fun goToStartEndList(view: View) {
 
-        val filteredGames = games.mapNotNull { it.state }
-        val pendingGamesCount = filteredGames.filter { it == Constants.pending }.size
-        val inProgressGamesCount = filteredGames.filter { it == Constants.inProgress }.size
-        val finishedGamesCount = filteredGames.filter { it == Constants.finished }.size
+        with(binding) {
+            when (view) {
+                floatingActionButtonStartList -> {
 
-        text_view_games_number.text = resources.getString(R.string.GAMES_NUMBER_TITLE, games.size)
-        button_pending.text_view_subtitle.text = "$pendingGamesCount"
-        button_in_progress.text_view_subtitle.text = "$inProgressGamesCount"
-        button_finished.text_view_subtitle.text = "$finishedGamesCount"
-    }
+                    recyclerViewGames.scrollToPosition(0)
+                    scrollPosition.set(ScrollPosition.TOP)
+                }
+                floatingActionButtonEndList -> {
 
-    private fun launchNotification(games: List<GameResponse>) {
-
-        val notifications = mutableMapOf<Int, Notification>()
-        var gameNames = ""
-        for (game in games) {
-
-            val intent = Intent(requireContext(), GameDetailActivity::class.java).apply {
-                putExtra("gameId", game.id)
-            }
-            val pendingIntent = PendingIntent.getActivity(requireContext(), game.id, intent, PendingIntent.FLAG_ONE_SHOT)
-
-            if (!sharedPrefHandler.notificationLaunched(game.id)) {
-
-                notifications[game.id] =
-                    NotificationCompat.Builder(requireContext(), Constants.channelId)
-                        .setSmallIcon(R.drawable.app_icon)
-                        .setContentTitle(resources.getString(R.string.NOTIFICATION_TITLE, game.name))
-                        .setContentText(resources.getString(R.string.NOTIFICATION_DESCRIPTION, Constants.dateToString(Date(), sharedPrefHandler), game.name))
-                        .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-                        .setContentIntent(pendingIntent)
-                        .setAutoCancel(true)
-                        .setCategory(NotificationCompat.CATEGORY_REMINDER)
-                        .setGroup(Constants.channelGroup)
-                        .build()
-                gameNames += game.name + ", "
+                    val position: Int = gamesAdapter.itemCount - 1
+                    recyclerViewGames.scrollToPosition(position)
+                    scrollPosition.set(ScrollPosition.END)
+                }
             }
         }
-        gameNames = gameNames.dropLast(2)
+    }
+    //endregion
 
+    //region Private methods
+    private fun initializeUI() {
 
-        val summaryNotification = NotificationCompat.Builder(requireContext(), Constants.channelId)
-            .setContentTitle(resources.getString(R.string.SUMMARY_NOTIFICATIONS_TITLE, games.size))
-            .setContentText(gameNames)
-            .setSmallIcon(R.drawable.app_icon)
-            .setStyle(
-                NotificationCompat.InboxStyle()
-                    .setBigContentTitle(resources.getString(R.string.SUMMARY_NOTIFICATIONS_TITLE, games.size))
-                    .setSummaryText(gameNames)
+        val application = activity?.application
+        viewModel = ViewModelProvider(
+            this,
+            GamesViewModelFactory(application)
+        ).get(GamesViewModel::class.java)
+        setupBindings()
+
+        with(binding) {
+
+            swipeRefreshLayout.apply {
+                isEnabled = this@GamesFragment.viewModel.swipeRefresh
+                setColorSchemeResources(R.color.colorPrimary)
+                setProgressBackgroundColorSchemeResource(R.color.colorSecondary)
+                setOnRefreshListener {
+                    this@GamesFragment.viewModel.loadGames()
+                }
+            }
+
+            gamesAdapter = GamesAdapter(
+                this@GamesFragment.viewModel.games.value ?: listOf(),
+                this@GamesFragment.viewModel.platforms,
+                null,
+                this@GamesFragment
             )
-            .setGroup(Constants.channelGroup)
-            .setGroupSummary(true)
-            .build()
+            recyclerViewGames.apply {
+                layoutManager = LinearLayoutManager(requireContext())
+                adapter = gamesAdapter
+                addOnScrollListener(object : RecyclerView.OnScrollListener() {
 
-        with(NotificationManagerCompat.from(requireContext())) {
-            for (notification in notifications) {
-                notify(notification.key, notification.value)
-                sharedPrefHandler.setNotificationLaunched(notification.key, true)
+                    override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
+                        super.onScrollStateChanged(recyclerView, newState)
+
+                        scrollPosition.set(
+                            if (!recyclerView.canScrollVertically(-1) && newState == RecyclerView.SCROLL_STATE_IDLE) {
+                                ScrollPosition.TOP
+                            } else if (!recyclerView.canScrollVertically(1) && newState == RecyclerView.SCROLL_STATE_IDLE) {
+                                ScrollPosition.END
+                            } else {
+                                ScrollPosition.MIDDLE
+                            }
+                        )
+                    }
+                })
             }
-            if (notifications.isNotEmpty()) notify(0, summaryNotification)
+
+            fragment = this@GamesFragment
+            viewModel = this@GamesFragment.viewModel
+            lifecycleOwner = this@GamesFragment
+            position = scrollPosition
         }
+
+        scrollPosition.set(ScrollPosition.TOP)
+    }
+
+    private fun setupBindings() {
+
+        viewModel.gamesLoading.observe(viewLifecycleOwner, { isLoading ->
+
+            enableStateButtons(!isLoading)
+            if (isLoading) {
+                showLoading()
+            } else {
+                hideLoading()
+            }
+        })
+
+        viewModel.gamesError.observe(viewLifecycleOwner, { error ->
+            manageError(error)
+        })
+
+        viewModel.games.observe(viewLifecycleOwner, {
+
+            val today = Constants.stringToDate(
+                Constants.dateToString(
+                    Date(),
+                    Constants.getDateFormatToShow(viewModel.language),
+                    viewModel.language
+                ),
+                Constants.getDateFormatToShow(viewModel.language),
+                viewModel.language
+            )
+            val gamesToNotify = ArrayList<GameResponse>()
+            for (game in it) {
+                if (game.releaseDate == today)
+                    gamesToNotify.add(game)
+            }
+            if (gamesToNotify.isNotEmpty()) launchNotification(gamesToNotify)
+        })
+
+        viewModel.gamesCount.observe(viewLifecycleOwner, {
+            setGamesCount(it)
+            setTitle(it.size)
+        })
     }
 
     private fun filter() {
@@ -358,92 +275,130 @@ class GamesFragment : BaseFragment(), GamesAdapter.OnItemClickListener, OnFilter
             ft.remove(prev)
         }
         ft.addToBackStack(null)
-        val dialogFragment = PopupFilterDialogFragment(currentFilters, this)
+        val dialogFragment = PopupFilterDialogFragment(viewModel.filters, this)
         dialogFragment.isCancelable = false
         dialogFragment.show(ft, "filterPopup")
     }
 
-    private fun add() {
-        launchActivity(GameDetailActivity::class.java)
+    private fun launchNotification(games: List<GameResponse>) {
+
+        val notifications = mutableMapOf<Int, Notification>()
+        var gameNames = Constants.EMPTY_VALUE
+        for (game in games) {
+
+            val intent = Intent(requireContext(), GameDetailActivity::class.java).apply {
+                putExtra(Constants.GAME_ID, game.id)
+            }
+            val pendingIntent = PendingIntent.getActivity(
+                requireContext(),
+                game.id,
+                intent,
+                PendingIntent.FLAG_ONE_SHOT
+            )
+
+            if (!viewModel.isNotificationLaunched(game.id)) {
+
+                notifications[game.id] =
+                    NotificationCompat.Builder(requireContext(), Notifications.CHANNEL_ID)
+                        .setSmallIcon(R.drawable.app_icon)
+                        .setContentTitle(
+                            resources.getString(
+                                R.string.notification_title,
+                                game.name
+                            )
+                        )
+                        .setContentText(
+                            resources.getString(
+                                R.string.notification_description,
+                                Constants.dateToString(
+                                    Date(),
+                                    Constants.getDateFormatToShow(viewModel.language),
+                                    viewModel.language
+                                ),
+                                game.name
+                            )
+                        )
+                        .setPriority(NotificationCompat.PRIORITY_DEFAULT)
+                        .setContentIntent(pendingIntent)
+                        .setAutoCancel(true)
+                        .setCategory(NotificationCompat.CATEGORY_REMINDER)
+                        .setGroup(Notifications.CHANNEL_GROUP)
+                        .build()
+                gameNames += game.name + ", "
+            }
+        }
+        gameNames = gameNames.dropLast(2)
+
+        val summaryNotification =
+            NotificationCompat.Builder(requireContext(), Notifications.CHANNEL_ID)
+                .setContentTitle(
+                    resources.getString(
+                        R.string.summary_notifications_title,
+                        games.size
+                    )
+                )
+                .setContentText(gameNames)
+                .setSmallIcon(R.drawable.app_icon)
+                .setStyle(
+                    NotificationCompat.InboxStyle()
+                        .setBigContentTitle(
+                            resources.getString(
+                                R.string.summary_notifications_title,
+                                games.size
+                            )
+                        )
+                        .setSummaryText(gameNames)
+                )
+                .setGroup(Notifications.CHANNEL_GROUP)
+                .setGroupSummary(true)
+                .build()
+
+        with(NotificationManagerCompat.from(requireContext())) {
+            for (notification in notifications) {
+
+                notify(notification.key, notification.value)
+                viewModel.setNotificationLaunched(notification.key, true)
+            }
+            if (notifications.isNotEmpty()) notify(0, summaryNotification)
+        }
     }
 
-    private fun sort() {
+    private fun setGamesCount(games: List<GameResponse>) {
 
-        val dialogView = LinearLayout(requireContext())
-        dialogView.orientation = LinearLayout.HORIZONTAL
+        val filteredGames = games.mapNotNull { it.state }
+        val pendingGamesCount = filteredGames.filter { it == State.PENDING_STATE }.size
+        val inProgressGamesCount = filteredGames.filter { it == State.IN_PROGRESS_STATE }.size
+        val finishedGamesCount = filteredGames.filter { it == State.FINISHED_STATE }.size
 
-        val ascendingPicker = getPicker(arrayOf(resources.getString(R.string.ASCENDING), resources.getString(R.string.DESCENDING)))
-        val ascendingPickerParams = LinearLayout.LayoutParams(
-            LinearLayout.LayoutParams.MATCH_PARENT,
-            LinearLayout.LayoutParams.WRAP_CONTENT
-        )
-        ascendingPickerParams.weight = 1f
-
-        val sortKeysPicker = getPicker(sortingValues)
-        val sortKeysPickerParams = LinearLayout.LayoutParams(
-            LinearLayout.LayoutParams.MATCH_PARENT,
-            LinearLayout.LayoutParams.WRAP_CONTENT
-        )
-        sortKeysPickerParams.weight = 1f
-
-        val params = LinearLayout.LayoutParams(50, 50)
-        params.gravity = Gravity.CENTER
-
-        dialogView.layoutParams = params
-        dialogView.addView(sortKeysPicker, sortKeysPickerParams)
-        dialogView.addView(ascendingPicker, ascendingPickerParams)
-
-        AlertDialog.Builder(context)
-            .setTitle(resources.getString(R.string.SORT_TITLE))
-            .setView(dialogView)
-            .setCancelable(false)
-            .setPositiveButton(resources.getString(R.string.ACCEPT)) { dialog, _ ->
-                sortKey = sortingKeys[sortKeysPicker.value]
-                sortAscending = ascendingPicker.value == 0
-                getContent(state, sortKey, sortAscending, currentFilters)
-                dialog.dismiss()
-            }
-            .setNegativeButton(resources.getString(R.string.CANCEL)) { dialog, _ ->
-                dialog.dismiss()
-            }
-            .show()
+        with(binding) {
+            buttonPending.text_view_subtitle.text = "$pendingGamesCount"
+            buttonInProgress.text_view_subtitle.text = "$inProgressGamesCount"
+            buttonFinished.text_view_subtitle.text = "$finishedGamesCount"
+        }
     }
 
-    private fun loadGames() {
+    private fun setTitle(gamesCount: Int) {
 
-        enableStateButtons(false)
-        gameAPIClient.getGames({
-
-            for (game in it) {
-                gameRepository.insertGame(game)
-            }
-            gameRepository.removeDisableContent(it)
-            val games = getContent(null, sortKey, sortAscending, currentFilters)
-            setGamesCount(games)
-            enableStateButtons(true)
-        }, {
-
-            manageError(it)
-            enableStateButtons(true)
-        })
+        val title = resources.getQuantityString(
+            R.plurals.games_number_title,
+            gamesCount,
+            Constants.getFormattedNumber(gamesCount)
+        )
+        (activity as AppCompatActivity?)?.supportActionBar?.title = title
     }
 
     private fun enableStateButtons(enable: Boolean) {
+        with(binding) {
 
-        swipe_refresh_layout.isRefreshing = !enable
-        button_pending.isEnabled = enable
-        button_in_progress.isEnabled = enable
-        button_finished.isEnabled = enable
+            swipeRefreshLayout.isRefreshing = !enable
+            buttonPending.isEnabled = enable
+            buttonInProgress.isEnabled = enable
+            buttonFinished.isEnabled = enable
+        }
     }
+    //endregion
+}
 
-    private fun getPicker(values: Array<String>): NumberPicker {
-
-        val picker = NumberPicker(requireContext())
-        picker.minValue = 0
-        picker.maxValue = values.size - 1
-        picker.wrapSelectorWheel = true
-        picker.descendantFocusability = NumberPicker.FOCUS_BLOCK_DESCENDANTS
-        picker.displayedValues = values
-        return picker
-    }
+enum class ScrollPosition {
+    TOP, MIDDLE, END, NONE
 }
