@@ -5,44 +5,57 @@ import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import es.upsa.mimo.gamercollection.models.*
-import es.upsa.mimo.gamercollection.models.base.BaseModel
-import es.upsa.mimo.gamercollection.persistence.daos.*
+import es.upsa.mimo.gamercollection.base.BaseModel
+import es.upsa.mimo.gamercollection.daos.*
+import es.upsa.mimo.gamercollection.models.responses.*
 import es.upsa.mimo.gamercollection.utils.Constants
 
-@Database(entities = [
-    FormatResponse::class,
-    GenreResponse::class,
-    PlatformResponse::class,
-    StateResponse::class,
-    GameResponse::class,
-    SongResponse::class,
-    SagaResponse::class], version = 1)
+@Database(
+    entities = [
+        FormatResponse::class,
+        GameResponse::class,
+        GenreResponse::class,
+        PlatformResponse::class,
+        SagaResponse::class,
+        SongResponse::class,
+        StateResponse::class], version = 1
+)
 @TypeConverters(ListConverter::class, DateConverter::class)
 abstract class AppDatabase : RoomDatabase() {
 
     abstract fun formatDao(): FormatDao
+    abstract fun gameDao(): GameDao
     abstract fun genreDao(): GenreDao
     abstract fun platformDao(): PlatformDao
-    abstract fun stateDao(): StateDao
-    abstract fun gameDao(): GameDao
     abstract fun sagaDao(): SagaDao
+    abstract fun stateDao(): StateDao
 
     companion object {
 
+        //region Private properties
         private var instance: AppDatabase? = null
+        //endregion
 
+        //region Public methods
         fun getAppDatabase(context: Context): AppDatabase {
 
             if (instance == null) {
                 synchronized(AppDatabase::class) {
-                    instance = Room.databaseBuilder(context.applicationContext, AppDatabase::class.java, Constants.databaseName).build()
+
+                    instance = Room.databaseBuilder(
+                        context.applicationContext,
+                        AppDatabase::class.java,
+                        Constants.DATABASE_NAME
+                    ).build()
                 }
             }
             return instance!!
         }
 
-        fun <T> getDisabledContent(currentValues: List<BaseModel<T>>, newValues: List<BaseModel<T>>): List<BaseModel<T>> {
+        fun <T> getDisabledContent(
+            currentValues: List<BaseModel<T>>,
+            newValues: List<BaseModel<T>>
+        ): List<BaseModel<T>> {
 
             val disabledContent = arrayListOf<BaseModel<T>>()
             for (currentValue in currentValues) {
@@ -53,5 +66,6 @@ abstract class AppDatabase : RoomDatabase() {
             }
             return disabledContent
         }
+        //endregion
     }
 }
