@@ -3,14 +3,17 @@ package es.upsa.mimo.gamercollection.persistence
 import androidx.room.TypeConverter
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
-import es.upsa.mimo.gamercollection.models.GameResponse
-import es.upsa.mimo.gamercollection.models.SongResponse
+import es.upsa.mimo.gamercollection.models.responses.GameResponse
+import es.upsa.mimo.gamercollection.models.responses.SongResponse
 import java.util.*
 
 class ListConverter {
 
+    //region Private properties
     private val gson = Gson()
+    //endregion
 
+    //region Public methods
     @TypeConverter
     fun stringToSongList(data: String?): List<SongResponse?>? {
         if (data == null) {
@@ -40,4 +43,5 @@ class ListConverter {
     fun gameListToString(games: List<GameResponse?>?): String? {
         return gson.toJson(games)
     }
+    //endregion
 }
