@@ -1,26 +1,46 @@
 package es.upsa.mimo.gamercollection.viewholders
 
-import android.content.Context
 import android.text.InputType
-import android.view.View
-import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import es.upsa.mimo.gamercollection.R
+import es.upsa.mimo.gamercollection.adapters.OnItemClickListener
+import es.upsa.mimo.gamercollection.databinding.SagaItemBinding
 import es.upsa.mimo.gamercollection.extensions.setReadOnly
-import es.upsa.mimo.gamercollection.models.SagaResponse
-import kotlinx.android.synthetic.main.saga_item.view.*
+import es.upsa.mimo.gamercollection.models.responses.SagaResponse
+import es.upsa.mimo.gamercollection.utils.Constants
 
-class SagasViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+class SagasViewHolder(private val binding: SagaItemBinding) :
+    RecyclerView.ViewHolder(binding.root) {
 
-    fun fillData(saga: SagaResponse, context: Context) {
+    //region Public methods
+    fun bind(saga: SagaResponse, onItemClickListener: OnItemClickListener) {
 
-        itemView.edit_text_name.setText(saga.name)
-        itemView.edit_text_name.setReadOnly(true, InputType.TYPE_NULL, 0)
-        itemView.image_view_arrow.setImageDrawable(ContextCompat.getDrawable(context, R.drawable.ic_keyboard_arrow_down_white_24dp))
-        itemView.image_view_arrow.visibility = if (saga.games.isNotEmpty()) View.VISIBLE else View.GONE
+        val gamesCount = saga.games.size
+        val title = if (gamesCount > 0) {
+            binding.root.context.resources.getQuantityString(
+                R.plurals.saga_title,
+                gamesCount,
+                saga.name,
+                gamesCount
+            )
+        } else {
+            saga.name
+        }
+
+        with(binding) {
+
+            editTextName.setText(title)
+            editTextName.setReadOnly(true, InputType.TYPE_NULL, 0)
+
+            darkMode = Constants.isDarkMode(root.context)
+            this.saga = saga
+            this.onItemClickListener = onItemClickListener
+        }
+
     }
 
     fun rotateArrow(value: Float) {
-        itemView.image_view_arrow.animate().setDuration(500).rotation(value).start()
+        binding.imageViewArrow.animate().setDuration(500).rotation(value).start()
     }
+    //endregion
 }
